requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
PyYAML>=6.0

# Bioinformatics
biopython>=1.79
mdtraj>=1.9.7
mdanalysis>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Structure analysis
prody>=2.0

# Deep learning (for BioEmu)
torch>=2.0.0

# CLI interface
click>=8.0.0
rich>=12.0.0
